Pushed v1.18 code
- Bugfix: Pasting copied sample data to an empty sample didn't work! Also fixed
  an issue with the sample length not changing when pasting data.
- Bugfix: Scopes would never stop showing looped samples after channel muting
- Bugfix: The sampling position line in the sampler screen would not behave
  correctly during "sample swapping".
- Bugfix: Left/right/up/down cursor keys should not be repeated in keyrepeat
  mode (toggled with Caps Lock).
- Windows bugfix: The Windows key could get stuck if you held down ALT while
  pressing it.
- Windows bugfix: Num Lock now works ike it should when the program is in focus
  (yes, this ruins drumpad mode, but it never worked right to begin with).
- The "real VU-meter" bars are now sinking a bit faster
- Code cleanup

+++ src/pt2_audio.h
@@ -40,9 +40,9 @@ typedef struct voice_t
 
 	// used for pt2_sync.c
 	uint8_t syncFlags;
-	int8_t syncVolume;
-	uint16_t syncPeriod;
-	uint16_t syncTriggerLength;
+	uint8_t syncVolume;
+	int32_t syncPeriod;
+	int32_t syncTriggerLength;
 	const int8_t *syncTriggerData;
 } paulaVoice_t;
